# implemented KDTree based sphere search for improved performance