Hey,

I modified your code such that the nearest neighbor search is now performed via a KDTree. The slight overhead of creating an additional data structure already pays off for pointclouds of ~10k points (`runExample()` with `n=10000, csRad=10` went down from 9 to 2.5sec and `runFixedExample()` with `n=10000, numSP=100, spSiye=10, csRad=10` went down from 13 to 2 sec)